Limitations under MIT license
he software and hardware is open source for experimentation / development and 
academic purposes only. This is supposed to be a not-for profit project
( commercialization is forbidden without the express written permission of Debargha Ganguly ).
In case of commercial distribution of the product, a modest percentage of the profits is 
to be paid to Debargha Ganguly for future development and maintainance purposes.